+++ SwiftMessages/MarginAdjustable.swift
@@ -30,3 +30,4 @@ public protocol MarginAdjustable {
     /// Safe area bottom adjustment in iOS 11+
     var safeAreaBottomOffset: CGFloat { get set }
 }
+