+++ SwiftMessages/PhysicsAnimation.swift
@@ -21,6 +21,7 @@ public class PhysicsAnimation: NSObject, Animator {
     public weak var delegate: AnimationDelegate?
     weak var messageView: UIView?
     weak var containerView: UIView?
+    var context: AnimationContext?
 
     public override init() {}
 
@@ -29,16 +30,19 @@ public class PhysicsAnimation: NSObject, Animator {
     }
 
     public func show(context: AnimationContext, completion: @escaping AnimationCompletion) {
+        NotificationCenter.default.addObserver(self, selector: #selector(adjustMargins), name: Notification.Name.UIDeviceOrientationDidChange, object: nil)
         install(context: context)
         showAnimation(context: context, completion: completion)
     }
 
     public func hide(context: AnimationContext, completion: @escaping AnimationCompletion) {
+        NotificationCenter.default.removeObserver(self)
         if panHandler?.isOffScreen ?? false {
             context.messageView.alpha = 0
             panHandler?.state?.stop()
         }
         let view = context.messageView
+        self.context = context
         CATransaction.begin()
         CATransaction.setCompletionBlock {
             view.alpha = 1
@@ -59,6 +63,7 @@ public class PhysicsAnimation: NSObject, Animator {
         let container = context.containerView
         messageView = view
         containerView = container
+        self.context = context
         view.translatesAutoresizingMaskIntoConstraints = false
         container.addSubview(view)
         switch placement {
@@ -69,36 +74,41 @@ public class PhysicsAnimation: NSObject, Animator {
         case .bottom:
             NSLayoutConstraint(item: container, attribute: .bottom, relatedBy: .equal, toItem: view, attribute: .bottom, multiplier: 1, constant: 0).isActive = true
         }
-        if let adjustable = view as? MarginAdjustable & UIView {
-            // Important to layout now in order to get the right safe area insets
-            container.layoutIfNeeded()
-            var top: CGFloat = 0
-            var bottom: CGFloat = 0
-            switch placement {
-            case .top:
-                top += adjustable.topAdjustment(container: container, context: context)
-            case .bottom:
-                bottom += adjustable.bottomAdjustment(container: container, context: context)
-            case .center:
-                break
-            }
-            adjustable.preservesSuperviewLayoutMargins = false
-            if #available(iOS 11, *) {
-                var margins = adjustable.directionalLayoutMargins
-                margins.top = top
-                margins.bottom = bottom
-                adjustable.directionalLayoutMargins = margins
-            } else {
-                var margins = adjustable.layoutMargins
-                margins.top = top
-                margins.bottom = bottom
-                adjustable.layoutMargins = margins
-            }
-        }
+        // Important to layout now in order to get the right safe area insets
+        container.layoutIfNeeded()
+        adjustMargins()
         NSLayoutConstraint(item: view, attribute: .leading, relatedBy: .equal, toItem: container, attribute: .leading, multiplier: 1, constant: 0).isActive = true
         NSLayoutConstraint(item: view, attribute: .trailing, relatedBy: .equal, toItem: container, attribute: .trailing, multiplier: 1, constant: 0).isActive = true
         container.layoutIfNeeded()
         installInteractive(context: context)
+    }
+
+    @objc public func adjustMargins() {
+        guard let adjustable = messageView as? MarginAdjustable & UIView,
+            let container = containerView,
+            let context = context else { return }
+        var top: CGFloat = 0
+        var bottom: CGFloat = 0
+        switch placement {
+        case .top:
+            top += adjustable.topAdjustment(container: container, context: context)
+        case .bottom:
+            bottom += adjustable.bottomAdjustment(container: container, context: context)
+        case .center:
+            break
+        }
+        adjustable.preservesSuperviewLayoutMargins = false
+        if #available(iOS 11, *) {
+            var margins = adjustable.directionalLayoutMargins
+            margins.top = top
+            margins.bottom = bottom
+            adjustable.directionalLayoutMargins = margins
+        } else {
+            var margins = adjustable.layoutMargins
+            margins.top = top
+            margins.bottom = bottom
+            adjustable.layoutMargins = margins
+        }
     }
 
     func showAnimation(context: AnimationContext, completion: @escaping AnimationCompletion) {