+++ SwiftMessages/WindowViewController.swift
@@ -50,10 +50,6 @@ class WindowViewController: UIViewController
     }
     
     override public var preferredStatusBarStyle: UIStatusBarStyle {
-        return config.preferredStatusBarStyle ?? UIApplication.shared.statusBarStyle
-    }
-    
-    override var prefersStatusBarHidden: Bool {
-        return UIApplication.shared.isStatusBarHidden
+        return config.preferredStatusBarStyle ?? super.preferredStatusBarStyle
     }
 }