+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 * Fix #128 move icons out of asset catalog to prevent mysterious crash
 * Fix #134 add support for `CenterAnimation` displayed on top or bottom instead of center (renamed to `PhysicsAnimation`).
+* Fix #129 adjust layout margins on orientation change to preserve layout when iOS hides status bar in landscape.
 
 
 ## 4.0.0