+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 4.0.1
+## 4.1.0
+
+### Features
+
+* Fix #134 add support for `CenterAnimation` displayed on top or bottom instead of center (renamed to `PhysicsAnimation`).
 
 ### Fixes
 
 * Fix #128 move icons out of asset catalog to prevent mysterious crash
-* Fix #134 add support for `CenterAnimation` displayed on top or bottom instead of center (renamed to `PhysicsAnimation`).
 * Fix #129 adjust layout margins on orientation change to preserve layout when iOS hides status bar in landscape.
+* Fix #131 by always completing hide/show animations if application isn't active.
 
 
 ## 4.0.0